+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## 2026-07-07 
 - Add ANOVA app and run_anova_app() function
+- Add first draft of linear regression app
 - Bump version to 0.0.2.9000
 
 ## 2026-07-03